Refactored Thermostat entities and Dto's

+++ IOT_DeviceManager.API/DTO/ThermostatDeviceMeasurementDto.cs
@@ -6,8 +6,8 @@ namespace IOT_DeviceManager.API.DTO
     public class ThermostatDeviceMeasurementDto : IDeviceMeasurementDto
     {
         public string Id { get; set; }
-        public float Temperature { get; set; }
-        public float SetPoint { get; set; }
+        public double Temperature { get; set; }
+        public double SetPoint { get; set; }
         public IDeviceStatusDto Status { get; set; }
         public DateTime TimeStamp { get; set; }
         public string DeviceId { get; set; }
